requirements.txt
# Core ML Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

# Data Science & Analysis
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
python-docx>=0.8.11
datasets>=2.12.0

# Jupyter & Notebook
jupyter>=1.0.0

# API and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
streamlit>=1.25.0
pydantic>=2.0.0

# Caching and Queue
redis>=4.5.0
celery>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
aiohttp>=3.8.0

# Utilities
psutil>=5.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
jinja2>=3.1.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0 